Moved around some file, classes and methods and changed added some error handling to the uppercaser function in utils
