+++ arrowhead_client/application/__init__.py
@@ -1,0 +1,1 @@
+from .app_core import ArrowheadApplication
