+++ arrowhead_client/application/core_services/service_registry.py
